requirements.txt
Flask==3.0.3
Flask_Cors==5.0.0
Jinja2==3.1.4
PyGithub==2.4.0
pytest==8.2.2
